feat: Added drawer and native navigation
