+++ src/App.tsx
@@ -1,23 +1,9 @@
-import { StatusBar } from "expo-status-bar";
-import { StyleSheet, Text, View } from "react-native";
 import { registerRootComponent } from "expo";
+import { NavigationContainer } from "./navigation/NavigationContainer";
+import "react-native-gesture-handler";
 
 export default function App() {
-  return (
-    <View style={styles.container}>
-      <Text>Start Here</Text>
-      <StatusBar style="auto" />
-    </View>
-  );
+  return <NavigationContainer />;
 }
 
 registerRootComponent(App);
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
